refactor(eventFormLogic): extract registration form helpers into pure logic
